+++ src/03-bot/NasdaqBot/Services/StubMarketService.cs
@@ -37,7 +37,7 @@ namespace NasdaqBot.Services
 
             return Task.FromResult<StockOrder>(new StockOrder
             {
-                Costs = rnd.NextDouble() * 40, OrderNumber = $"Order{rnd.Next(100, 999)}", OrderType = "Buy",
+                Costs = (decimal)rnd.NextDouble() * 40, OrderNumber = $"Order{rnd.Next(100, 999)}", OrderType = "Buy",
                 Amount = request.Amount, Limit = request.OrderLimit, 
                 StockSymbol = request.StockSymbol
             });
